refactor: extract useAnalyzerStatus hook for improved data handling in DemoStatusRow
